Update to user interface
Modifications made to the CREATE action for new cities.
