+++ assets/scripts/auth/api.js
@@ -43,58 +43,9 @@ const changePassword = (data) => {
   })
 }
 
-const createCity = (name, state, country) => {
-  console.log('new city created. data is:', name, state, country)
-  return $.ajax({
-    url: app.host + '/cities',
-    method: 'POST',
-    headers: {
-      Authorization: 'Token token=' + store.user.token
-    },
-    data: {
-      'city': {
-        'name': name,
-        'state': state,
-        'country': country
-      }
-    }
-  })
-}
-
-const getCities = function () {
-  return $.ajax({
-    url: app.host + '/cities',
-    method: 'GET',
-    headers: {
-      Authorization: 'Token token=' + store.user.token
-    }
-  })
-}
-
-// const updateCity = (name, state, country) => {
-//   console.log('update city to server.')
-//   return $.ajax({
-//     url: config.apiOrigin + '/cities/' + city.id,
-//     method: 'PATCH',
-//     headers: {
-//       Authorization: 'Token token=' + store.user.token
-//     },
-//     data: {
-//       'city': {
-//         'name': name,
-//         'state': state,
-//         'country': country
-//       }
-//     }
-//   })
-// }
-
 module.exports = {
   signUp,
   signIn,
   signOut,
-  changePassword,
-  createCity,
-  getCities
-  // updateCity
+  changePassword
 }
