+++ index.html
@@ -38,7 +38,8 @@
   <div class="create-city-message"></div>
 
 <!-- City input fields -->
-<div class="form-group" id="city-input-fields">
+<!-- <div class="form-group" id="city-input-fields">
+<form id="city-form">
 <div class="form-group">
   <label for="city-input" class="col-2 col-form-label">City</label>
   <div class="col-10">
@@ -57,15 +58,43 @@
     <input class="form-control" type="text" id="country-input">
   </div>
 </div>
+<input type="submit" name="submit" value="Add City!">
+</form>
 <button type="button" class="add-city-btn btn-primary">Add</button>
-</div>
+</div> -->
+<div class="btn-container">
+    <button type="button" class="new-city-button btn btn-primary" data-toggle="modal" data-target="#addCity-modal">
+      Add City</button>
+  </div>
 
-<div>
+  <!-- Add City Modal -->
+  <div class="modal fade" id="addCity-modal" tabindex="-1" role="dialog" aria-labelledby="myModalLabel">
+    <div class="modal-dialog" role="document">
+      <div class="modal-content">
+        <div class="modal-header">
+          <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
+          <h4 class="modal-title" id="sign-up-modal-label">Add City</h4>
+        </div>
+        <div class="modal-body">
+          <form id="city-form">
+            <div class="form-group">
+              <input type="text" name="city[name]" id="city-input" placeholder="City (required)">
+              <input type="text" name="city[state]" id="state-input" placeholder="State">
+              <input type="text" name="city[country]" id="country-input" placeholder="Country">
+              <input type="submit" name="submit" value="Add city">
+            </div>
+          </form>
+        </div>
+      </div>
+    </div>
+  </div>
+
+<div id="view-cities">
 <button type="button" class="view-cities-btn btn-primary">View Cities</button>
 </div>
 
 <!-- list of cities populated using handlebars-->
-<div id="content" class="content">list should be here</div>
+<div id="content" class="content"></div>
 
   <!-- Sign Up Modal -->
   <div class="modal fade" id="signUp-modal" tabindex="-1" role="dialog" aria-labelledby="myModalLabel">
